Refactor and update various components for improved functionality and compatibility
- Added `@OptIn(ExperimentalCoroutinesApi::class)` annotation to `DownloadUtil` and `HomeViewModel` for experimental coroutine usage.
- Simplified cache handling in `CachePlaylistViewModel` by removing unnecessary null checks.
- Updated import statements from `hilt.navigation.compose.hiltViewModel` to `hilt.lifecycle.viewmodel.compose.hiltViewModel` across multiple screens for consistency with the latest Hilt library.
- Replaced `Divider` with `HorizontalDivider` in `EnhancedPreferenceEntry` for better UI consistency.
- Improved handling of cached static data in `MusicService` by removing redundant null checks.
- Adjusted `ShowMediaInfo` to directly use nullable types for view counts and likes, enhancing readability.
- Updated `ContentSettings` to use the new `ExposedDropdownMenuAnchorType` for dropdown menus.
- Various minor adjustments and optimizations across UI screens and view models to enhance performance and maintainability.

+++ app/src/main/kotlin/com/anitail/music/MainActivity.kt
@@ -113,7 +113,7 @@ import androidx.core.content.ContextCompat
 import androidx.core.net.toUri
 import androidx.core.util.Consumer
 import androidx.core.view.WindowCompat
-import androidx.hilt.navigation.compose.hiltViewModel
+import androidx.hilt.lifecycle.viewmodel.compose.hiltViewModel
 import androidx.lifecycle.lifecycleScope
 import androidx.media3.common.MediaItem
 import androidx.media3.common.Player
@@ -387,7 +387,7 @@ class MainActivity : AppCompatActivity() {
                     while (isActive) {
                         playerConnection?.service?.lanJamServer?.clientList?.let { clients ->
                             jamViewModel.updateActiveConnections(
-                                clients.map { client -> client.ip to client.connectedAt.toLong() }
+                                clients.map { client -> client.ip to client.connectedAt }
                             )
                         }
                         // Incrementar delay para reducir carga del CPU
